% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{vulkan_available}
\alias{vulkan_available}
\title{Check Vulkan AV1 availability}
\usage{
vulkan_available()
}
\value{
\code{TRUE} if the package was compiled with Vulkan AV1 encode support.
}
\description{
Check Vulkan AV1 availability
}
