% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{vulkan_devices}
\alias{vulkan_devices}
\title{List Vulkan-capable GPU devices}
\usage{
vulkan_devices()
}
\value{
Character vector of device names. Devices supporting
  \code{VK_KHR_VIDEO_ENCODE_AV1} are marked with \code{[AV1]}.
}
\description{
List Vulkan-capable GPU devices
}
